+++ KFXAdditions/Classes/UIApplication+KFXAdditions.m
@@ -85,6 +85,20 @@
 
 }
 
+
+- (BOOL)kfx_pushNotificationsEnabled {
+    
+    if ([[UIApplication sharedApplication] respondsToSelector:@selector(currentUserNotificationSettings)]) {
+        UIUserNotificationType types = [[[UIApplication sharedApplication] currentUserNotificationSettings] types];
+        return (types & UIUserNotificationTypeAlert);
+    }
+    else {
+        UIRemoteNotificationType types = [[UIApplication sharedApplication] enabledRemoteNotificationTypes];
+        return (types & UIRemoteNotificationTypeAlert);
+    }
+}
+
+
 @end